Enhanced the navigation for new pets and med entries
The new functionality allows someone to create a new pet and immediately add a medication. The entry will only save if they hit SAVE. I removed the auto save from Add as this was confusing.

I changed SAVE to DONE on the Med Entry page to avoid confustion, as the meds only save when the user hits SAVE on the detail view.

Some UI clean up also in here.